Recursion
For this assignment, you are expected to write an interactive program which, according to the user input, creates
dynamic responses and do simple calculations.
A compound object will be defined interactively and you program will keep track of a simple property and return a
result based on the composition of the object.